minor fix layout
