+++ lib/view/pages/course_history_view.dart
@@ -3,6 +3,8 @@ import 'package:presencee/view/widgets/diagram_history.dart';
 import 'package:presencee/view/widgets/header_history.dart';
 import 'package:presencee/view/widgets/list_history.dart';
 
+import '../../theme/constant.dart';
+
 class CourseHistory extends StatefulWidget {
   const CourseHistory({Key? key}) : super(key: key);
 
@@ -14,16 +16,52 @@ class _CourseHistoryState extends State<CourseHistory> {
   @override
   Widget build(BuildContext context) {
 
-    return const DefaultTabController(
+    return DefaultTabController(
       length: 2,
       child: Scaffold(
-        appBar: PreferredSize(
+        appBar: const PreferredSize(
           preferredSize: Size.fromHeight(120),
           child: HeaderHistory()),
-        body: TabBarView(
+        body: Column(
           children: [
-            ListHistory(),
-            DiagramHistory()
+            const SizedBox(height: 37),
+            Flexible(
+              flex: 1,
+              child: Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 25),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Text(
+                      'Bahasa Indonesia (MU)',
+                      style: AppTextStyle.poppinsTextStyle(
+                        color: AppTheme.black,
+                        fontsWeight: FontWeight.w600,
+                        fontSize: 24
+                      )
+                    ),
+                    Text(
+                      'Siswandi',
+                      style: AppTextStyle.poppinsTextStyle(
+                        color: AppTheme.black,
+                        fontsWeight: FontWeight.w400,
+                        fontSize: 20
+                      )
+                    ),
+                  ],
+                ),
+              )
+            ),
+            const Flexible(
+              flex: 8,
+              child: TabBarView(
+                children: [
+                  ListHistory(),
+                  DiagramHistory()
+                ],
+              ),
+            )
           ],
         ),
       ),
